trading_bot/requirements.txt
python-binance
python-dotenv
fastapi
uvicorn
jinja2
python-multipart
